// packages/phoenix-theme/src/themeColors.d.ts
export type ThemeBackgroundColors = {
  backgroundCardOnflyoutDefaultDisabled: string;
  backgroundCardOnflyoutDefaultHover: string;
  backgroundCardOnflyoutDefaultPressed: string;
  backgroundCardOnflyoutDefaultRest: string;
  backgroundCardOnprimaryAltDisabled: string;
  backgroundCardOnprimaryAltHover: string;
  backgroundCardOnprimaryAltPressed: string;
  backgroundCardOnprimaryAltRest: string;
  backgroundCardOnprimaryDefaultDisabled: string;
  backgroundCardOnprimaryDefaultHover: string;
  backgroundCardOnprimaryDefaultPressed: string;
  backgroundCardOnprimaryDefaultRest: string;
  backgroundCardOnprimaryDefaultSelected: string;
  backgroundCardOnsecondaryAltDisabled: string;
  backgroundCardOnsecondaryAltHover: string;
  backgroundCardOnsecondaryAltPressed: string;
  backgroundCardOnsecondaryAltRest: string;
  backgroundCardOnsecondaryDefaultDisabled: string;
  backgroundCardOnsecondaryDefaultHover: string;
  backgroundCardOnsecondaryDefaultPressed: string;
  backgroundCardOnsecondaryDefaultRest: string;
  backgroundControlProgressBarEmptyRest: string;
  backgroundCtrlActivebrandDisabled: string;
  backgroundCtrlActivebrandHover: string;
  backgroundCtrlActivebrandPressed: string;
  backgroundCtrlActivebrandRest: string;
  backgroundCtrlBrandDisabled: string;
  backgroundCtrlBrandHover: string;
  backgroundCtrlBrandPressed: string;
  backgroundCtrlBrandRest: string;
  backgroundCtrlNeutralDisabled: string;
  backgroundCtrlNeutralHover: string;
  backgroundCtrlNeutralPressed: string;
  backgroundCtrlNeutralRest: string;
  backgroundCtrlOnimageHover: string;
  backgroundCtrlOnimagePressed: string;
  backgroundCtrlOnimageRest: string;
  backgroundCtrlOutlineDisabled: string;
  backgroundCtrlOutlineHover: string;
  backgroundCtrlOutlinePressed: string;
  backgroundCtrlOutlineRest: string;
  backgroundCtrlShapesafeActivebrandDisabled: string;
  backgroundCtrlShapesafeActivebrandRest: string;
  backgroundCtrlShapesafeNeutralDisabled: string;
  backgroundCtrlShapesafeNeutralHover: string;
  backgroundCtrlShapesafeNeutralPressed: string;
  backgroundCtrlShapesafeNeutralRest: string;
  backgroundCtrlSubtleDisabled: string;
  backgroundCtrlSubtleHover: string;
  backgroundCtrlSubtleHoverSplit: string;
  backgroundCtrlSubtlePressed: string;
  backgroundCtrlSubtleRest: string;
  backgroundFlyoutColorblend: string;
  backgroundFlyoutLumblend: string;
  backgroundFlyoutSolid: string;
  backgroundLayerPrimarySolid: string;
  backgroundLayerPrimaryStop1: string;
  backgroundLayerPrimaryStop2: string;
  backgroundLayerPrimaryStop3: string;
  backgroundLayerSecondary: string;
  backgroundLayerTertiary: string;
  backgroundSmoke: string;
  backgroundToolbar: string;
  backgroundWebPagePrimary: string;
  backgroundWebPageSecondary: string;
  backgroundWindowPrimaryColorBlend: string;
  backgroundWindowPrimaryLumBlend: string;
  backgroundWindowPrimarySolid: string;
  backgroundWindowSecondaryColorBlend: string;
  backgroundWindowSecondaryLumBlend: string;
  backgroundWindowSecondarySolid: string;
  backgroundWindowTabBandColorBlend: string;
  backgroundWindowTabBandInactive: string;
  backgroundWindowTabBandLumBlend: string;
  backgroundWindowTabBandSolid: string;
  backgroundControlBrandRest: string;
  backgroundControlInputRest: string;
  backgroundControlNeutralHover: string;
  backgroundControlNeutralRest: string;
  backgroundControlSubtleDisabled: string;
  backgroundControlSubtleHover: string;
  backgroundControlSubtlePressed: string;
  backgroundControlSubtleRest: string;
};
// ...
export type ThemeForegroundColors = {
  foregroundControlBrandRest: string;
  foregroundControlIconOnSubtleDisabled: string;
  foregroundControlIconOnSubtleHover: string;
  foregroundControlIconOnSubtlePressed: string;
  foregroundControlIconOnSubtleRest: string;
  foregroundControlNeutralPrimaryRest: string;
  foregroundControlNeutralSecondaryDisabled: string;
  foregroundControlNeutralSecondaryHover: string;
  foregroundControlNeutralSecondaryPressed: string;
  foregroundControlNeutralSecondaryRest: string;
  foregroundControlOnActiveBrandRest: string;
  foregroundControlShapeSafeNeutralHover: string;
  foregroundControlShapeSafeNeutralPressed: string;
  foregroundContentBrandPrimary: string;
  foregroundContentNeutralPrimary: string;
  foregroundContentNeutralSecondary: string;
  foregroundControlLinkBrandRest: string;
  foregroundControlOnBrandRest: string;
  foregroundControlShapesafeNeutralHover: string;
  foregroundControlShapesafeNeutralPressed: string;
  foregroundControlShapesafeNeutralRest: string;
  foregroundCtrlActivebrandDisabled: string;
  foregroundCtrlActivebrandHover: string;
  foregroundCtrlActivebrandPressed: string;
  foregroundCtrlActivebrandRest: string;
  foregroundCtrlBrandDisabled: string;
  foregroundCtrlBrandHover: string;
  foregroundCtrlBrandPressed: string;
  foregroundCtrlBrandRest: string;
  foregroundCtrlHintDefault: string;
  foregroundCtrlIconOnneutralDisabled: string;
  foregroundCtrlIconOnneutralHover: string;
  foregroundCtrlIconOnneutralPressed: string;
  foregroundCtrlIconOnneutralRest: string;
  foregroundCtrlIconOnoutlineDisabled: string;
  foregroundCtrlIconOnoutlineHover: string;
  foregroundCtrlIconOnoutlinePressed: string;
  foregroundCtrlIconOnoutlineRest: string;
  foregroundCtrlIconOnsubtleDisabled: string;
  foregroundCtrlIconOnsubtleHover: string;
  foregroundCtrlIconOnsubtlePressed: string;
  foregroundCtrlIconOnsubtleRest: string;
  foregroundCtrlNeutralPrimaryDisabled: string;
  foregroundCtrlNeutralPrimaryHover: string;
  foregroundCtrlNeutralPrimaryPressed: string;
  foregroundCtrlNeutralPrimaryRest: string;
  foregroundCtrlNeutralSecondaryDisabled: string;
  foregroundCtrlNeutralSecondaryHover: string;
  foregroundCtrlNeutralSecondaryPressed: string;
  foregroundCtrlNeutralSecondaryRest: string;
  foregroundCtrlOnactivebrandDisabled: string;
  foregroundCtrlOnactivebrandHover: string;
  foregroundCtrlOnactivebrandPressed: string;
  foregroundCtrlOnactivebrandRest: string;
  foregroundCtrlOnbrandDisabled: string;
  foregroundCtrlOnbrandHover: string;
  foregroundCtrlOnbrandPressed: string;
  foregroundCtrlOnbrandRest: string;
  foregroundCtrlOnimageRest: string;
  foregroundCtrlOnoutlineDisabled: string;
  foregroundCtrlOnoutlineHover: string;
  foregroundCtrlOnoutlinePressed: string;
  foregroundCtrlOnoutlineRest: string;
  foregroundCtrlOnsubtleDisabled: string;
  foregroundCtrlOnsubtleHover: string;
  foregroundCtrlOnsubtlePressed: string;
  foregroundCtrlOnsubtleRest: string;
  foregroundCtrlOntransparentDisabled: string;
  foregroundCtrlOntransparentHover: string;
  foregroundCtrlOntransparentPressed: string;
  foregroundCtrlOntransparentRest: string;
  foregroundNeutralDisabled: string;
  foregroundNeutralSecondary: string;
};

export type ThemeShadowColors = {
  shadowToolbarKeyColor: string;
  shadowToolbarAmbientColor: string;
  shadowWindowActiveKeyColor: string;
  shadowWindowActiveAmbientColor: string;
  shadowWindowInactiveKeyColor: string;
  shadowWindowInactiveAmbientColor: string;
  shadowCardRestKeyColor: string;
  shadowCardRestAmbientColor: string;
  shadowCardHoverKeyColor: string;
  shadowCardPressedKeyColor: string;
  shadowCardDisabledKeyColor: string;
  shadowFlyoutKeyColor: string;
  shadowFlyoutAmbientColor: string;
  shadowCtrlOnDragKeyColor: string;
  shadowCtrlOnDragAmbientColor: string;
  shadowLayerKeyColor: string;
  shadowLayerAmbientColor: string;
};

export type ThemeMaterialColors = {
  materialAcrylicDefaultSolid: string;
  materialAcrylicDefaultColorblend: string;
  materialAcrylicDefaultLumblend: string;
  materialAcrylicShellDefaultSolid: string;
  materialAcrylicShellDefaultColorblend: string;
  materialAcrylicShellDefaultLumblend: string;
  materialMicaThinSolid: string;
  materialMicaThinColorblend: string;
  materialMicaThinLumblend: string;
  materialMicaDarkerSolid: string;
  materialMicaDarkerColorblend: string;
  materialMicaDarkerLumblend: string;
  materialMicaDefaultSolid: string;
  materialMicaDefaultColorblend: string;
  materialMicaDefaultLumblend: string;
};

export type ThemeStatusColors = {
  statusBrandBackground: string;
  statusBrandForeground: string;
  statusBrandStroke: string;
  statusBrandTintBackground: string;
  statusBrandTintForeground: string;
  statusBrandTintStroke: string;
  statusDangerBackground: string;
  statusDangerForeground: string;
  statusDangerStroke: string;
  statusDangerTintBackground: string;
  statusDangerTintForeground: string;
  statusDangerTintStroke: string;
  statusSuccessBackground: string;
  statusSuccessForeground: string;
  statusSuccessStroke: string;
  statusSuccessTintBackground: string;
  statusSuccessTintForeground: string;
  statusSuccessTintStroke: string;
  statusWarningBackground: string;
  statusWarningForeground: string;
  statusWarningStroke: string;
  statusWarningTintBackground: string;
  statusWarningTintForeground: string;
  statusWarningTintStroke: string;
  statusInformativeBackground: string;
  statusInformativeForeground: string;
  statusInformativeStroke: string;
  statusInformativeTintBackground: string;
  statusInformativeTintForeground: string;
  statusInformativeTintStroke: string;
  statusImportantBackground: string;
  statusImportantForeground: string;
  statusImportantStroke: string;
  statusImportantTintBackground: string;
  statusImportantTintForeground: string;
  statusImportantTintStroke: string;
  statusAwayForeground: string;
  statusOofForeground: string;
};

export type ThemeNullColors = {
  nullColor: string;
  nullColorBackground: string;
  nullColorBackgroundHover: string;
  nullColorBackgroundPressed: string;
  nullColorStroke: string;
};

export type ThemeAiColors = {
  aiShimmerStop1: string;
  aiShimmerStop2: string;
  aiShimmerStop3: string;
  aiShimmerStop4: string;
  aiBrandStop1: string;
  aiBrandStop2: string;
  aiBrandStop3: string;
  aiBrandStop4: string;
};